[refactor] 파일 위치 옮김
