+++ resource-server/src/main/java/com/inhabas/api/controller/ContestBoardController.java
@@ -1,6 +1,6 @@
 package com.inhabas.api.controller;
 
-import com.inhabas.api.argumentResolver.Authenticated;
+import com.inhabas.api.web.argumentResolver.Authenticated;
 import com.inhabas.api.domain.member.MemberId;
 import com.inhabas.api.dto.contest.DetailContestBoardDto;
 import com.inhabas.api.dto.contest.ListContestBoardDto;
